requirements_freeze.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anywidget==0.9.21
arrow==1.4.0
asciitree==0.3.3
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
bids-validator-deno==2.4.1
bidsschematools==1.2.1
blessed==1.32.0
blosc2==4.1.1
boto3==1.42.59
botocore==1.42.59
brainglobe-atlasapi==2.3.0
brainglobe-heatmap==0.6.0
brainglobe-space==1.0.3
brainglobe-utils==0.11.1
brainrender==2.1.18
bs4==0.0.2
cached-property==2.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.4.0
click==8.1.8
click-didyoumean==0.3.1
cloudpickle==3.1.2
colorama==0.4.6
colorcet==3.1.0
colorlog==6.10.1
comm==0.2.3
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dandi==0.74.3
dandischema==0.12.1
dask==2026.1.2
debugpy==1.8.20
decorator==5.2.1
deepcomparer==0.4.0
deprecated==1.3.1
deprecation==2.1.0
dnspython==2.8.0
docstring-parser==0.17.0
email-validator==2.3.0
etelemetry==0.3.1
executing==2.2.1
fasteners==0.20
figrid==0.1.7
flake8==7.3.0
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fscacher==0.4.4
fsspec==2026.2.0
globus-sdk==4.4.1
google==3.0.0
graphviz==0.21
h5py==3.15.1
hdmf==5.0.0
hdmf-zarr==0.12.0
humanize==4.15.0
ibl-neuropixel==1.10.0
ibl-style==0.1.0
-e file:///home/heberto/development/ibl_conversion/IBL-to-nwb
iblatlas==0.10.0
ibllib @ git+https://github.com/int-brain-lab/ibllib@e89a339a57f8a67c68fe8bac1c148af97d9e6dad
iblqt==0.8.2
iblutil==1.20.0
idna==3.11
imagecodecs==2026.1.14
imageio==2.37.2
iniconfig==2.3.0
interleave==0.3.0
ipykernel==7.2.0
ipython==9.10.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
isodate==0.7.2
isoduration==20.11.0
jaraco-classes==3.4.0
jaraco-context==6.1.0
jaraco-functools==4.4.0
jedi==0.19.2
jeepney==0.9.0
jmespath==1.1.0
joblib==1.5.3
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.8.0
jupyter-core==5.9.1
jupyterlab-widgets==3.0.16
k3d==2.17.0
keyring==25.7.0
keyrings-alt==5.0.2
kiwisolver==1.4.9
lazy-loader==0.4
llvmlite==0.46.0
locket==1.0.0
loguru==0.7.3
markdown-it-py==4.0.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
meshio==5.3.5
ml-dtypes==0.5.4
more-itertools==10.8.0
morphapi==0.3.2
morphio==3.4.0
msgpack==1.1.2
mtscomp==1.0.2
multidict==6.7.1
myterial==1.2.1
natsort==8.4.0
ndindex==1.10.1
ndx-anatomical-localization==0.1.0
ndx-events==0.2.2
ndx-ibl @ git+https://github.com/catalystneuro/ndx-ibl@231169415fb15287bfeaac7622758f4ac9bbee88
ndx-pose==0.2.2
neo==0.14.3
nest-asyncio==1.6.0
networkx==3.6.1
neuroconv @ git+https://github.com/catalystneuro/neuroconv@4b6454edec1ccce41d5fdb24e61555f27802430c
neurom==4.0.4
nibabel==5.4.0
nptdms==1.10.0
numba==0.64.0
numcodecs==0.15.1
numexpr==2.14.1
numpy==2.2.0
nwb-video-widgets==0.1.5
nwbinspector==0.7.0
one-api==3.4.2
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==2.3.3
parse==1.21.1
parso==0.8.6
partd==1.4.2
patsy==1.0.2
pexpect==4.9.0
phylib==2.7.0
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
pooch==1.9.0
probeinterface==0.3.1
prompt-toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
psychofit==1.0.0.post0
psygnal==0.15.1
ptyprocess==0.7.0
pure-eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==23.0.1
pycodestyle==2.14.0
pycparser==3.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pyflakes==3.4.0
pygments==2.19.2
pyinspect==0.1.0
pyjwt==2.11.0
pynapple==0.10.3
pynrrd==1.1.3
pynwb==3.1.3
pyout==0.8.1
pyparsing==3.3.2
pyqt5==5.15.11
pyqt5-qt5==5.15.18
pyqt5-sip==12.18.0
pyqtgraph==0.14.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyyaml==6.0.3
pyzmq==27.1.0
qt-niu==0.1.0
qtpy==2.4.3
quantities==0.16.4
rapidyaml==0.10.0
referencing==0.37.0
remfile==0.1.13
requests==2.32.5
responses==0.26.0
retry==0.9.2
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==14.3.3
rpds-py==0.30.0
ruamel-yaml==0.19.1
ruff==0.15.4
s3transfer==0.16.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
secretstorage==3.5.0
semantic-version==2.10.0
shapely==2.1.2
six==1.17.0
slidingrp==1.1.1
slurmio==0.1.2
soupsieve==2.8.3
sparse==0.18.0
spikeinterface==0.103.2
stack-data==0.6.3
statsmodels==0.14.6
superqt==0.7.8
tables==3.11.1
tabulate==0.9.0
tenacity==9.1.4
tensorstore==0.1.81
threadpoolctl==3.6.0
tifffile==2026.2.24
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
traittypes==0.2.3
treelib==1.8.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
vedo==2026.6.1
vtk==9.6.0
wcwidth==0.6.0
webcolors==25.10.0
widgetsnbextension==4.0.15
wrapt==2.1.1
xarray==2026.2.0
yarl==1.23.0
zarr==2.18.7
zarr-checksum==0.4.7
